backend/requirements/base.txt
Django>=4.2,<4.3
djangorestframework>=3.14,<3.15
djangorestframework-simplejwt>=5.3,<5.4
psycopg2-binary>=2.9,<3.0
python-dotenv>=1.0,<2.0
django-cors-headers>=4.3,<4.4
celery>=5.3,<5.4
redis>=5.0,<6.0
requests>=2.31,<3.0
Pillow>=10.1,<11.0
django-filter>=23.3
drf-nested-routers>=0.93.4
urllib3<2.0
docker==5.0.3
requests==2.31.0
